[integration] Add Apache License headers
I’ve also added it to the CI integration workflow; it made more sense than to add this step into the integration `tox.ini` file since we copy the other Python packages into the directory (which we’d then end up scanning). We could rework the tox file a bit, but adding it to the CI workflow was just simpler.