[all] Add license classifier
Some systems automatically scan for package license information to “approve of” packages. A PyPI classifier is the only way to designate the license.

There’s a `license` keyword arg for the `setup()` function; however, that is for when the project’s license is not in the list of classifiers (https://pypi.org/pypi?%3Aaction=list_classifiers). (ref: https://docs.python.org/3.6/distutils/setupscript.html#additional-meta-data)