[integration] Fix bubbling up of errors
Previously, integration tests would only fail if the last command in `integration/tox_commands.sh` would run (either the running of the job or running of the integration test suite if there is one) would fail. If unit tests or job running (if there were also an integration test suite) would fail, it would not be caught.