[docs] Workaround for docs build issue
The newly added `sphinx-reredirects` extension has a bug in it. I’ve made a PR
(https://gitlab.com/documatt/sphinx-reredirects/-/merge_requests/1) but who
knows if/when it’ll get addressed. This work around runs the same “linkcheck”
command but checks the error message, and succeeds instead of errors if this
bug is encountered (this is a benign bug otherwise).

Revert this commit (and bump the version of `sphinx-reredirects`) if/when this
bug gets addressed.