[exec] Add  DirectGKERunner support (#197)
* [exec] Implement new direct-on-GKE runner

* [exec] Add input message acking when Direct on GKE

When running direct runner on GKE, instead of auto-acking when a message is
immediately read from the Pub/Sub queue, we ack when the message is considered
done. A message is done & ack’ed when:
* the pipeline successfully ran, but before any write to event output happens if any,
* the message is skipped because output data already exists,
* the message is dropped because input data does not exist, or
* the message is dropped because it was not the intended recipient.

* [exec] Add unit tests for the following classes in the
pubsub_message_manager module:

* PubSubKlioMessage
* MessageManager

* [exec] Update string name of direct-on-GKE runner

* [exec] don't flatten to_pass_thru if it's None (no data inputs)

* [exec] use unique labels for KlioAckInputMessage txfm

* [exec] Log benign exc if pubsub queue is empty for ~20+ seconds

* [exec] Add unit tests for the following classes in the
pubsub_message_manager module:

* PubSubKlioMessage
* MessageManager

* [exec] Add warning on missing entity_id mapping and tests for KlioAckInput message transform

* [exec] Add test for KlioPubsubReadEvaluator

* [exec] Add basic GkeDirectRunner tests

* [exec] Add docstrings to pubsub_message_manager

* [exec] Add docstring for GkeDirectRunner

* [exec] Fix for when pulling pubsub raises a DeadlineExceeded exc

We're catching a benign error if it's been a while since a
message has been available to be pulled off the queue.
But when this happens, the variable is undefined, raising an
UnboundedLocalError when trying to continue on.

* [lib] Fix metric name for timeouts

I noticed this when running a test pipeline trying to emit metrics for a Grafana dashboard: every other metric around KlioMessage handling is prefixed with `kmsg-`.

* [exec] Move runner unit tests to correct dir level

* [exec,lib] Move pubsub message manager to klio

* [exec] Rename module and update tests

evaluators.py originally had all the pubsub message manager logic in it, but we’ve
now moved that into klio. Minor code changes were made - instead of parsing
the KlioMessage within the evaluator, that is now handled by the pubsub
message manager.

* [exec] Update exec runner to use helper AckInputMessage transform

* [exec] Limit google-cloud-pubsub dep

This is already limited by `apache-beam[gcp]` but when running tox, the latest
(unsupported) version of google-cloud-pubsub was pulled in.

* [lib,exec] Use a threadpoolexecutor rather than threads

For the managing of pubsub messages

* [lib] Reuse metrics dispatcher thdpool exec

The metrics dispatcher may be initialized repeatedly (as seen when running on
DirectGKERunner). This ensures that we re-use an existing threadpool executor
rather than creating new ones.

* Small fixups from PR feedback

* [IT] Fix weird log timing issue when there's threading

* [lib] Fix patching for pubsub ack testing

* [lib] Increase default pubsub manager sleep time to 10s

So that the debug logs on not extending a message is not as annoying

Co-authored-by: Lynn Root <lynn@spotify.com>
Co-authored-by: Dan Simon <dsimon@spotify.com>