[lib] Decrease sleep between possible msg extensions
If the pubsub manager sleeps for 10 seconds per loop, the default extension
is 30 seconds, and the manager only extends if it’s been longer than 24 seconds
(30 * 80%), by the time the manager tries to extend the message deadline,
it could have already expired. Decreasing the sleep between loops should
provide 1-2 attempts to extend the deadline before it’s too late.